ic_cdk
Module api

System API and low level functions for it.

Modules

call

APIs to make and manage calls in the canister.

management_canister

Functions and types for calling the IC management canister.

stable

APIs to manage stable memory.

Functions

call_context_instruction_counter

Returns the number of WebAssembly instructions the canister has executed within the call context of the current Message execution since Call context creation.

caller

Returns the caller of the current call.

canister_balance

Gets the amount of funds available in the canister.

canister_balance128

Gets the amount of funds available in the canister.

canister_version

Gets the value of canister version.

cycles_burn

Burns cycles from the canister.

data_certificate

When called from a query call, returns the data certificate authenticating certified_data set by this canister.

id

Returns the canister id as a blob.

in_replicated_execution

Checks if in replicated execution.

instruction_counter

Returns the number of instructions that the canister executed since the last entry point.

is_controller

Determines if a Principal is a controller of the canister.

performance_counter

Gets the value of specified performance counter.

print

Prints the given message.

set_certified_data

Sets the certified data of this canister.

set_global_timer

Sets global timer.

time

Gets current timestamp, in nanoseconds since the epoch (1970-01-01)

trap

Traps with the given message.