ic_cdk::api
Module call

APIs to make and manage calls in the canister.

Structs

ArgDecoderConfig

Config to control the behavior of decoding canister endpoint arguments.

CallReplyWriter

An io::Write for message replies.

ManualReply

Pretends to have the Candid type T, but unconditionally errors when serialized.

Enums

RejectionCode

Rejection code from calling another canister.

Functions

accept_message

Accepts the ingress message.

arg_data

Returns the argument data in the current call. Traps if the data cannot be decoded.

arg_data_raw

Returns the argument data as bytes.

arg_data_raw_size

Gets the len of the raw-argument-data-bytes.

call

Performs an asynchronous call to another canister.

call_raw

Performs an asynchronous call to another canister and pay cycles at the same time.

call_raw128

Performs an asynchronous call to another canister and pay cycles (in u128) at the same time.

call_with_config

Performs an asynchronous call to another canister and pay cycles (in u128). It also allows setting a quota for decoding the return values. The decoding quota is strongly recommended when calling third-party or untrusted canisters.

call_with_payment

Performs an asynchronous call to another canister and pay cycles at the same time.

call_with_payment128

Performs an asynchronous call to another canister and pay cycles (in u128) at the same time.

is_recovering_from_trap

Tells you whether the current async fn is being canceled due to a trap/panic.

method_name

Returns the name of current canister method.

msg_cycles_accept

Moves cycles from the call to the canister balance.

msg_cycles_accept128

Moves cycles from the call to the canister balance.

msg_cycles_available

Returns the amount of cycles that were transferred by the caller of the current call, and is still available in this message.

msg_cycles_available128

Returns the amount of cycles that were transferred by the caller of the current call, and is still available in this message.

msg_cycles_refunded

Returns the amount of cycles that came back with the response as a refund.

msg_cycles_refunded128

Returns the amount of cycles that came back with the response as a refund.

notify

Like notify_with_payment128, but sets the payment to zero.

notify_raw

Like notify, but sends the argument as raw bytes, skipping Candid serialization.

notify_with_payment128

Sends a one-way message with payment cycles attached to it that invokes method with arguments args on the principal identified by id, ignoring the reply.

performance_counterDeprecated

Gets the value of specified performance counter

reject

Rejects the current call with the message.

reject_code

Returns the rejection code for the call.

reject_message

Returns the rejection message.

reply

Replies to the current call with a candid argument.

reply_raw

Replies with the bytes passed

result

Returns a result that maps over the call

Type Aliases

CallResult

The result of a Call.