ic_cdk::api
Module management_canister

Functions and types for calling the IC management canister.

This module is a direct translation from the interface description.

The functions and types defined in this module serves these purposes:

Make it easy to construct correct request data.

Handle the response ergonomically.

For those calls require cycles payments, the cycles amount is an explicit argument.

Modules

bitcoin

The IC Bitcoin API.

ecdsa

Threshold ECDSA signing API.

http_request

Canister HTTP request.

main

The main functionalities in the IC management canister.

provisional

Provisional functions only available in local development instances.

schnorr

Threshold Schnorr signing API.