ic_cdk::api::management_canister
Module ecdsa

Threshold ECDSA signing API.

Structs

EcdsaKeyId

ECDSA KeyId.

EcdsaPublicKeyArgument

Argument type of ecdsa_public_key.

EcdsaPublicKeyResponse

Response Type of ecdsa_public_key.

SignWithEcdsaArgument

Argument type of sign_with_ecdsa.

SignWithEcdsaResponse

Response type of sign_with_ecdsa.

Enums

EcdsaCurve

ECDSA Curve.

Functions

ecdsa_public_key

Return a SEC1 encoded ECDSA public key for the given canister using the given derivation path.

sign_with_ecdsa

Return a new ECDSA signature of the given message_hash that can be separately verified against a derived ECDSA public key.