ic_cdk::api::management_canister
Module main

The main functionalities in the IC management canister.

Most of the functions are for managing canister lifecycle. raw_rand is also included in this module.

Structs

CanisterChange

Represents a canister change as stored in the canister history.

CanisterIdRecord

A wrapper of canister id.

CanisterInfoRequest

Argument type of canister_info.

CanisterInfoResponse

Return type of canister_info.

CanisterSettings

Canister settings.

CanisterStatusResponse

Return type of canister_status.

ChunkHash

Return type of upload_chunk and stored_chunks.

ClearChunkStoreArgument

Argument type of clear_chunk_store.

CodeDeploymentRecord

Details about a canister code deployment.

ControllersChangeRecord

Details about updating canister controllers.

CreateCanisterArgument

Argument type of create_canister.

CreationRecord

Details about a canister creation.

DefiniteCanisterSettings

Like CanisterSettings.

DeleteCanisterSnapshotArgs

Argument type of delete_canister_snapshot.

FromCanisterRecord

Details about a canister change initiated by a canister (called originator).

FromUserRecord

Details about a canister change initiated by a user.

InstallChunkedCodeArgument

Argument type of install_chunked_code.

InstallCodeArgument

Argument type of install_code.

LoadCanisterSnapshotArgs

Argument type of load_canister_snapshot.

LoadSnapshotRecord

Details about loading canister snapshot.

QueryStats

Query statistics, returned by canister_status.

Snapshot

A snapshot of the state of the canister at a given point in time.

StoredChunksArgument

Argument type of stored_chunks.

SubnetInfoArgs

Argument type of subnet_info.

SubnetInfoResult

Result type of subnet_info.

TakeCanisterSnapshotArgs

Argument type of take_canister_snapshot.

UpdateSettingsArgument

Argument type of update_settings.

UpgradeFlags

Flags for canister installation with CanisterInstallMode::Upgrade.

UploadChunkArgument

Argument type of update_chunk.

Enums

CanisterChangeDetails

Provides details on the respective canister change.

CanisterChangeOrigin

Provides details on who initiated a canister change.

CanisterInstallMode

The mode with which a canister is installed.

CanisterStatusType

Status of a canister.

CodeDeploymentMode

The mode with which a canister is installed.

LogVisibility

todo

WasmPersistenceMode

WASM memory persistence setting for UpgradeFlags.

Functions

canister_info

Get public information about the canister.

canister_status

Get status information about the canister.

clear_chunk_store

See IC method clear_chunk_store.

create_canister

Register a new canister and get its canister id.

delete_canister

Delete a canister from the IC.

delete_canister_snapshot

Delete a specified snapshot that belongs to an existing canister.

deposit_cycles

Deposit cycles into the specified canister.

install_chunked_code

Install code into a canister where the code has previously been uploaded in chunks.

install_code

Install code into a canister.

list_canister_snapshots

List the snapshots of the canister.

load_canister_snapshot

Load a snapshot onto the canister.

raw_rand

Get 32 pseudo-random bytes.

start_canister

Start a canister if the canister status was stopped or stopping.

stop_canister

Stop a canister.

stored_chunks

See IC method stored_chunks.

subnet_info

Get subnet metadata.

take_canister_snapshot

Take a snapshot of the specified canister.

uninstall_code

Remove a canister’s code and state, making the canister empty again.

update_settings

Update the settings of a canister.

upload_chunk

See IC method upload_chunk.

Type Aliases

CanisterId

Canister ID is Principal.

SnapshotId

ID of a canister snapshot.

WasmModule

WASM module.