Module ic_cdk::api::management_canister::main

The main functionalities in the IC management canister.

Most of the functions are for managing canister lifecycle. raw_rand is also included in this module.

Structs

CanisterIdRecord

A wrapper of canister id.

CanisterSettings

Canister settings.

CanisterStatusResponse

Argument type of canister_status.

CreateCanisterArgument

Argument type of create_canister.

DefiniteCanisterSettings

Like CanisterSettings.

InstallCodeArgument

Argument type of install_code.

UpdateSettingsArgument

Argument type of update_settings.

Enums

CanisterInstallMode

The mode with which a canister is installed.

CanisterStatusType

Status of a canister.

Constants

CREATE_CANISTER_CYCLES

Cycles cost to create a canister.

Functions

canister_status

Get status information about the canister.

create_canister

Register a new canister and get its canister id.

create_canister_with_extra_cycles

create_canister and specify extra cycles to the new canister.

delete_canister

Delete a canister from the IC.

deposit_cycles

Deposit cycles into the specified canister.

install_code

Install code into a canister.

raw_rand

Get 32 pseudo-random bytes.

start_canister

Start a canister if the canister status was stopped or stopping.

stop_canister

Stop a canister.

uninstall_code

Remove a canister’s code and state, making the canister empty again.

update_settings

Update the settings of a canister.

Type Definitions

CanisterId

Canister ID is Principal.

WasmModule

WASM module.